requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
python-dotenv>=1.0.1
pydantic>=2.6.4
langchain>=0.1.13
langchain-groq>=0.1.1
langchain-community>=0.0.29
langgraph>=0.0.30
pinecone-client>=3.2.0
langchain-pinecone>=0.0.3
sentence-transformers>=2.6.0
pypdf>=4.1.0
